Update SW_SC.py
Modified SW_SC.py to avoid the generation of a large number of redundant files when the user is interested in multiple thresholds. Now, only a single threshold can be specified, but a manifest will be generated that documents the proportion of Xmer coverage prior to adding each peptide. This manifest can be used to down select from the full peptides to hit different Xmer coverage thresholds.

Also added a new "-i" option that can be used to specify the input target fasta files. In this case, the user just provides a string (with wildcards) to provide to glob in order to specify the input files. This is a workaround for the long commands generated when there are a lot of input files and the wildcard is expanded by the shell prior to issuing the command.